+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 
 ## 0.0.45
 
+- Support `lbl` widget as the non-tile-themed version of `label` (i.e. `::TkLabel` not `::Tk::Tile::TLabel`)
 - Add `#proxy` method to all Tk widgets to return Glimmer widget proxy objects (e.g. `Tk::Tile::TButton#proxy` returns `Glimmer::Tk::WidgetProxy` object)
 - Provide question mark alias of `event.drop_accepted` (i.e. `drop_accepted?`)
 - Have `DragAndDropEvent` `source`/`target` be the enhanced Glimmer widget proxy instead of the unenhanced Tk widget
+- Look into improving code that uses `TkLabel` explicitly in Hello, Drag and Drop! (do `event.tooltip.content {label {...} }` with Glimmer DSL instead)
 - Fix issue with dropping button and list unto checkbox in Hello, Drag and Drop!
 
 ## 0.0.44